Add Successful Login Test Case
Created a new spec file for all positive test cases.

The describe statement in the cypress/e2e/login_positive_spec.cy.js outlies the steps needed to login. It firstly creates a LoginPage object which allows us to visit the website and locate all the fields and buttons we require for this action. It then sets variables for email and passwords. Once these are set it feeds them to the loginPage object to perform the login action which is outlines in the cypress/pages/loginPage.js file

Signed-off-by: Mohammad Azwad Rashid <m.azwad.rashid@gmail.com>