Add Assertion statements to validate login
Two assertion statements have been added to verify that the login was successful.

The first one checks for a change in URL. Once successfully logged in the URL should end in /new_library/t/320073 for the user credential provided.

The second assertion checks the presense of the user icon on the top right of the page that allows the user to navigate to the user's settings and profile.

Signed-off-by: Mohammad Azwad Rashid <m.azwad.rashid@gmail.com>